+++ ftc/sim_parallel.py
@@ -64,3 +64,22 @@ def evaluate(N, threshold=0.5):
         alt_errors = np.append(alt_errors, info["alt_error"])
     recovery_rate = calculate_recovery_rate(alt_errors, threshold=threshold)
     print(f"Recovery rate is {recovery_rate:.3f}.")
+
+
+def evaluate_pos(threshold=np.ones(3)):
+    data = fym.load("data.h5")["env"]
+    errors = (data["posd"] - data["plant"]["pos"]).squeeze()
+    error_norms = np.linalg.norm(errors, axis=0)
+    print(f"Position trajectory error norms are {error_norms}.")
+    return np.all(error_norms <= threshold)
+
+
+def evaluate_mfa(eval, verbose=False):
+    data = fym.load("data.h5")["env"]
+    mfa = np.all(data["mfa"])
+    if mfa == eval:
+        if verbose:
+            print("MFA Success")
+    else:
+        if verbose:
+            print("MFA Fails")
